requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0
requests==2.31.0
twilio==8.11.0
python-telegram-bot==20.7
apscheduler==3.10.4
sqlalchemy==2.0.23
python-dateutil==2.8.2
pytz==2023.3
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
